refactor: move some munsell conversion code to the conversions file
